working on calculator WIP
Work stopped at 13:30 to switch context to career
today skimmed javascript ninja function chapters to clarify how and why
to use the different ways of declaring javascript functions and why the
anonymous function is used so heavily in javascript.

set up qunit testing for this project.
working on test driven development for this project.